refactor: move `TSPoint`
