+++ query_server/query/src/extension/expr/aggregate_function/mod.rs
@@ -4,12 +4,15 @@ mod gauge;
 mod sample;
 mod state_agg;
 
+use std::sync::Arc;
+
 use datafusion::arrow::array::ArrayRef;
-use datafusion::arrow::datatypes::DataType;
+use datafusion::arrow::datatypes::{DataType, Field, Fields};
 use datafusion::common::Result as DFResult;
+use datafusion::error::DataFusionError;
 use datafusion::scalar::ScalarValue;
 use spi::query::function::FunctionMetadataManager;
-use spi::Result;
+use spi::{QueryError, Result};
 
 pub const SAMPLE_UDAF_NAME: &str = "sample";
 pub const COMPACT_STATE_AGG_UDAF_NAME: &str = "compact_state_agg";
@@ -25,16 +28,101 @@ pub fn register_udafs(func_manager: &mut dyn FunctionMetadataManager) -> Result<
     Ok(())
 }
 
-trait AggResult {
+pub trait AggResult {
     fn to_scalar(self) -> DFResult<ScalarValue>;
 }
 
-trait AggState: Sized {
+pub trait AggState: Sized {
     fn try_to_state(&self) -> DFResult<Vec<ScalarValue>>;
     fn try_from_arrays(
         input_data_types: &[DataType],
         arrays: &[ArrayRef],
     ) -> DFResult<Option<Self>>;
+}
+
+#[derive(Debug, Clone, PartialEq)]
+pub struct TSPoint {
+    ts: ScalarValue,
+    val: ScalarValue,
+}
+
+impl TSPoint {
+    pub fn try_new_null(time_data_type: DataType, value_data_type: DataType) -> DFResult<Self> {
+        let ts = ScalarValue::try_from(time_data_type)?;
+        let val = ScalarValue::try_from(value_data_type)?;
+        Ok(Self { ts, val })
+    }
+
+    pub fn val(&self) -> &ScalarValue {
+        &self.val
+    }
+
+    pub fn ts(&self) -> &ScalarValue {
+        &self.ts
+    }
+}
+
+impl AggResult for TSPoint {
+    fn to_scalar(self) -> DFResult<ScalarValue> {
+        let TSPoint { ts, val } = self;
+        let ts_data_type = ts.get_datatype();
+        let val_data_type = val.get_datatype();
+
+        Ok(ScalarValue::Struct(
+            Some(vec![ts, val]),
+            Fields::from([
+                Arc::new(Field::new("ts", ts_data_type, true)),
+                Arc::new(Field::new("val", val_data_type, true)),
+            ]),
+        ))
+    }
+}
+
+impl TSPoint {
+    pub fn try_from_scalar(scalar: ScalarValue) -> DFResult<Self> {
+        match scalar {
+            ScalarValue::Struct(None, fields) => {
+                debug_assert!(fields.len() == 2, "Expected 2 fields, got {:?}", fields);
+
+                let time_data_type = fields[0].data_type();
+                let value_data_type = fields[1].data_type();
+                let ts = ScalarValue::try_from(time_data_type)?;
+                let val = ScalarValue::try_from(value_data_type)?;
+                Ok(Self { ts, val })
+            }
+            ScalarValue::Struct(Some(vals), _) => {
+                let ts = vals[0].clone();
+                let val = vals[1].clone();
+                Ok(Self { ts, val })
+            }
+            _ => Err(DataFusionError::External(Box::new(QueryError::Internal {
+                reason: format!("Expected struct, got {:?}", scalar),
+            }))),
+        }
+    }
+
+    pub fn data_type(&self) -> DFResult<DataType> {
+        Ok(DataType::Struct(Fields::from([
+            Arc::new(Field::new("ts", self.ts.get_datatype(), true)),
+            Arc::new(Field::new("val", self.val.get_datatype(), true)),
+        ])))
+    }
+}
+
+fn scalar_to_points(value: ScalarValue) -> DFResult<Vec<TSPoint>> {
+    match value {
+        ScalarValue::List(Some(vals), _) => {
+            let points = vals
+                .into_iter()
+                .map(TSPoint::try_from_scalar)
+                .collect::<DFResult<Vec<_>>>()?;
+            Ok(points)
+        }
+        ScalarValue::List(None, _) => Ok(vec![]),
+        _ => Err(DataFusionError::External(Box::new(QueryError::Internal {
+            reason: format!("Expected list, got {:?}", value),
+        }))),
+    }
 }
 
 #[cfg(test)]
